backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]
beanie==1.22.6
python-jose==3.3.0
cryptography==41.0.4
pydantic~=2.4.2
motor~=3.3.1
pymongo~=4.5.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6